pyproject.toml
[build-system]
requires = ["flit_core >= 3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "aiidalab-qe-muon"
description="AiiDAlab QE plugin for muon spectoscopies."
version = "1.0.0"
authors = [
    {name = "Miki Bonacci", email = "miki.bonacci@psi.ch"},
]
readme = "README.md"
license = {file = "LICENSE.txt"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
]
keywords = ["aiidalab", "plugin","muon"]
requires-python = ">=3.8"
dependencies = [
    "aiida-muon @ git+https://github.com/positivemuon/aiida-muon.git@v1.0.2",
    "pymatgen>=2023.9.25",
    "aiida-workgraph"
]

[project.optional-dependencies]
docs = [
    "sphinx",
    "docutils",
    "sphinx-copybutton~=0.3.0",
    "sphinx-book-theme~=0.1.0",
    "sphinx-click~=2.7.1",
]

pre-commit = [
    "pre-commit>=2.21.0",
]

tests = [
    "pgtest~=1.3",
    "pytest~=6.0",
    "pytest-regressions~=2.3",
]

[tool.ruff.lint]
ignore = ["E501", "E402", "TRY003", "RUF012", "N806"]

[project.urls]
Source = "https://github.com/mikibonacci/aiidalab-qe-vibroscopy"

[project.entry-points.'aiidalab_qe.properties']
"muonic" = "aiidalab_qe_muon.app:property"

[project.entry-points.'aiida.workflows']
"muon_app.implant_muon" = "aiidalab_qe_muon.workflows.implantmuonworkchain:ImplantMuonWorkChain"

[project.scripts]
"install_muon_codes" = "aiidalab_qe_muon.scripts.post_install:InstallCodes"